Re-enable python unit tests
